system: merge 'event' and 'thread' module
* seems like there is no more reason to maintain separately
* renamed test files for windows
